split up between image extraction and label extraction
